ref #17 move folder tests to folder charts
