+++ frontend/src/containers/ChartRanking/__test__/ChartRanking.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { useSelector, useDispatch } from 'react-redux';
-import ChartRanking from '../containers/ChartRanking/ChartRanking';
+import ChartRanking from '../ChartRanking';
 window.React = React;
 
 jest.mock('react-redux', () => ({
